introducing AMServiceRequest helper util - which provides methods to get the data from the server
This can be used to fire off the service request on App startup while the rest of the UI is loading. This way, once the data is available from the server, it can be used right away by the UI instead of making the user wait.

Updated TempleListActivity to use this interface